lights work; buzzer works-ish; button not done
Lights show green when key on hook.
Buzzer works for tested scenarios. Buzzer sounds when key taken off hook after delay and for a set duration. It will turn off if key placed on hook during delay or duration. if another key off hook after, buzzer mechanism repeats. Does not matter which hook key is placed on (don't know if we want to add that). 
Button not implemented.